third_party/blink/web_tests/fast/scrolling/events/scrollend-event-fired-to-element-with-overscroll-behavior.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="resources/scroll-util.js"></script>
<style>
#overscrollXDiv {
  width: 600px;
  height: 600px;
  overflow: scroll;
  overscroll-behavior-x: contain;
}
#overscrollYDiv {
  width: 500px;
  height: 500px;
  overflow: scroll;
  overscroll-behavior-y: none;
}
#targetDiv {
  width: 400px;
  height: 400px;
  overflow: scroll;
}
.content {
  width:800px;
  height:800px;
}
</style>

<body style="margin:0">
<div id="overscrollXDiv">
  <div class=content>
    <div id="overscrollYDiv">
      <div class=content>
        <div id="targetDiv">
          <div class="content">
          </div>
        </div>
      </div>
    </div>
  </div>
</div>
</body>

<script>
if (window.internals)
  internals.runtimeFlags.overscrollCustomizationEnabled = true;

var target_div = document.getElementById('targetDiv');
var rect = target_div.getBoundingClientRect();
var x = rect.left + rect.width / 2;
var y = rect.top + rect.height / 2;

var horizontal_scrollend_arrived = false;
var vertical_scrollend_arrived = false;
function onHorizontalScrollEnd(event) {
  assert_false(event.cancelable);
  assert_false(event.bubbles);
  horizontal_scrollend_arrived = true;
}
function onVerticalScrollEnd(event) {
  assert_false(event.cancelable);
  assert_false(event.bubbles);
  vertical_scrollend_arrived = true;
}
// Test that both "onscrollend" and addEventListener("scrollend"... works.
document.getElementById('overscrollXDiv').onscrollend = onHorizontalScrollEnd;
document.getElementById('overscrollYDiv').
    addEventListener("scrollend", onVerticalScrollEnd);

async function test_scrollend_event_fire(source_device,
    precise_scrolling_delta) {
  horizontal_scrollend_arrived = false;
  vertical_scrollend_arrived = false;
  await waitForCompositorCommit();

  // Scroll left on target div and wait for the element with overscroll-x to
  // get scrollend event.
  await scrollLeft(source_device, precise_scrolling_delta);
  await waitFor(() => { return horizontal_scrollend_arrived; },
      'Expected element did not receive scrollend event after scroll left ' +
      'on target.');
  assert_equals(target_div.scrollLeft, 0);

  // Scroll up on target div and wait for the element with overscroll-y to get
  // scrollend event.
  await scrollUp(source_device, precise_scrolling_delta);
  await waitFor(() => { return vertical_scrollend_arrived; },
      'Expected element did not receive scrollend event after scroll up on ' +
      'target.');
  assert_equals(target_div.scrollTop, 0);
}

promise_test (async (t) => {
  // Make sure that no scrollend event is sent to document or target_div.
  document.addEventListener("scrollend",
      t.unreached_func("Document got unexpected scrollend event."));
  target_div.addEventListener("scrollend",
      t.unreached_func("target_div got unexpected scrollend event."));
  return test_scrollend_event_fire(GestureSourceType.TOUCH_INPUT);
}, 'Tests that the last element in the cut scroll chain gets scrollend ' +
   'event when no element scrolls by touch.');

promise_test (async (t) => {
  // Make sure that no scrollend event is sent to document or target_div.
  document.addEventListener("scrollend",
      t.unreached_func("Document got unexpected scrollend event."));
  target_div.addEventListener("scrollend",
      t.unreached_func("target_div got unexpected scrollend event."));
  return test_scrollend_event_fire(GestureSourceType.MOUSE_INPUT,
    true /* precise_scrolling_delta */);
}, 'Tests that the last element in the cut scroll chain gets scrollend ' +
   'event when no element scrolls by wheel in non-animated scroll path.');

promise_test (async (t) => {
  // Make sure that no scrollend event is sent to document or target_div.
  document.addEventListener("scrollend",
      t.unreached_func("Document got unexpected scrollend event."));
  target_div.addEventListener("scrollend",
      t.unreached_func("target_div got unexpected scrollend event."));
  return test_scrollend_event_fire(GestureSourceType.MOUSE_INPUT,
    false /* precise_scrolling_delta */);
}, 'Tests that the last element in the cut scroll chain gets scrollend ' +
   'event when no element scrolls by wheel in smooth scrolling path.');
</script>
